git-svn-id: svn+ssh://src.apple.com/svn/cups/easysw/current@901 a1ca3aef-8c08-0410-bb20-df032aa958be

+++ scheduler/cups-driverd.cxx
@@ -76,7 +76,7 @@ static const char * const ppd_types[] =	/* ppd-type values */
 typedef struct				/**** PPD record ****/
 {
   time_t	mtime;			/* Modification time */
-  size_t	size;			/* Size in bytes */
+  off_t		size;			/* Size in bytes */
   int		model_number;		/* cupsModelNumber */
   int		type;			/* ppd-type */
   char		filename[512],		/* Filename */
